pyproject.toml
[project]
dynamic = ["version"]
name = "chtc-logs-tool"
dependencies = [
    "typer",
    "pytz",
    "dotenv",
    "python-magic"
]

requires-python = ">=3.11"  # Fixed missing closing quote

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
chtc-log-tools = "log_tools.cli:app"
